Speed up feasibility-weighted LCB acquisition
If feasibility estimation is moderately expensive, optimizing without
analytical gradients incurs a significant performance penalty.
However, this is not sufficient as the gradients of the acquisition functions
themselves require the estimated values, in which case the total evaluation count
might even exceed the overhead caused by finite differencing.

These issues can be alleviated by computing the value of the acquisition function
and its gradient jointly, in which case we can reuse estimator and LCB computations.